Address review feedback on PR #16
Both codepulse findings verified and correct. This PR moved the BF16 cyclic/paged
cache to symmetric BF16 K/V but left the benchmark and the published contracts still
saying FP16 V -- which is the same asymmetry the PR exists to remove, just in the
places a compiler cannot check.

- bench/ops/context_kv_materialize_bench.cu bound the cyclic cache V plane as
  DType::FP16 while K was BF16. The Op validates both planes against
  d256_kv_cache_profile, so every nonzero case failed validation. Now BF16, and the
  bench runs: 79.9 us at width=1 batch=1, real timings across the sweep.

- Four published contracts and two maintainer documents advertised FP16 V:
  ops/context_kv_materialize.h, ops/sliding_window_attention.h,
  ops/softmax_attention.h (the paged BF16 context), docs/maintainer/paged-kv-cache.md
  and docs/maintainer/qwen3.8-27b-dflash2.md. All now state BF16 K and V.

Ground truth for all of it is d256_kv_cache_profile, which returns {BF16, BF16, ...}
for the BFloat16 storage, and plane_types.h static_asserts that the value plane is
BF16 "symmetric with K". A stale contract is how this contamination keeps coming back
on merges -- a reader trusts the header, writes __half, and it is byte-compatible.

Full suite 123/123.

+++ bench/ops/context_kv_materialize_bench.cu
@@ -155,7 +155,9 @@ struct Fixture {
                 CyclicKVCacheLayerView{
                      .k        = Tensor(cache_k[static_cast<std::size_t>(layer)].p, DType::BF16,
                                         {kHeadDim, kCapacity, kHeads, kLaneCapacity}),
-                     .v        = Tensor(cache_v[static_cast<std::size_t>(layer)].p, DType::FP16,
+                     // BF16, symmetric with K. The Op validates both planes against
+                     // d256_kv_cache_profile, so an FP16 V here fails every nonzero case.
+                     .v        = Tensor(cache_v[static_cast<std::size_t>(layer)].p, DType::BF16,
                                         {kHeadDim, kCapacity, kHeads, kLaneCapacity}),
                      .capacity = kCapacity,
                      .padded_capacity = kCapacity,
